fix component lib importing
