+++ dan_layer/template_macros/src/lib.rs
@@ -21,7 +21,6 @@
 //  USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 mod ast;
-mod component;
 mod template;
 
 use proc_macro::TokenStream;
